requirements.txt
pycryptodome==3.20.0
python-decouple==3.8
streamlit==1.36.0
twilio==9.1.1
